Raise delete and cut to Canvas
